+++ ChangeLog.md
@@ -1,3 +1,14 @@
+3.0.3
+=====
+
+### Significant changes relative to 3.0.2:
+
+1. Fixed an issue in the Windows TurboVNC Viewer whereby a left Alt key press,
+followed by a left Alt key release, caused the keyboard focus to be redirected
+to the system menu, and subsequent keystrokes were consumed by the system menu
+until left Alt or Esc was pressed to dismiss the menu.
+
+
 3.0.2
 =====